// src/modules/vo/UserInfoVo.d.ts
/**
 * UmsUser
 */
export interface UmsUser {
    active?: boolean;
    alias?: string;
    avatar?: string;
    bio?: string;
    createTime?: Date;
    email?: string;
    id?: string;
    modifyTime?: Date;
    phone?: string;
    roleId?: number;
    score?: number;
    status?: boolean;
    username?: string;
    openid?: string;
    unionid?: string;
}